Change keyword default type to Int for platform stability reasons

+++ src/ConstructiveSolidGeometry/VolumePrimitives/Box.jl
@@ -74,8 +74,8 @@ function Box(::Type{CO}=ClosedPrimitive;
     hX = 1,
     hY = 1,
     hZ = 1,
-    origin = zero(CartesianPoint{Int64}), 
-    rotation = one(SMatrix{3, 3, Int64, 9})
+    origin = zero(CartesianPoint{Int}), 
+    rotation = one(SMatrix{3, 3, Int, 9})
 ) where {CO}
     Box(CO, hX, hY, hZ, origin, rotation)
 end
